Refac: Compile Time Calculation & New Interface.
1. New Technical Index Available.
2. New Interface, CompileTime Calculation
